+++ app/lib/util-service.js
@@ -647,7 +647,7 @@ angular.module('lizard-nxt')
   })();
 
   this.MIN_TIME = (new Date("Jan 01, 1900")).getTime();
-  this.MAX_TIME = (new Date()).getTime() + 24 * 60 * 60 * 1000;
+  this.MAX_TIME = (new Date()).getTime() + 20 * 24 * 60 * 60 * 1000; // 20 Days
   this.TIMELINE_LEFT_MARGIN = 60;
   this.TIMELINE_RIGHT_MARGIN = 40;
   this.OMNIBOX_WIDTH = 420;
@@ -670,8 +670,8 @@ angular.module('lizard-nxt')
     return leftMargin;
   };
 
-  this.getCurrentWidth = function () {
-    return window.innerWidth - (
+  this.getCurrentWidth = function (element) {
+    return element[0].clientWidth - (
       this.TIMELINE_LEFT_MARGIN + this.TIMELINE_RIGHT_MARGIN
     );
   };